move record_queries
